requirements.txt
gdal==3.12.2
libgdal-netcdf==3.12.2
libgdal-jp2openjpeg==3.12.2
rasterio==1.5.0 
numpy==2.4.2
pandas==3.0.0 ß
matplotlib==3.10.8  
progress==1.6.1  
requests==2.32.5
pyyaml==6.0.3
keras==3.13.2
tensorflow==2.20.0 (pip)
joblib==1.5.3
psutil==7.2.2
scikit-learn==1.8.0 

wget
s3cmd